requirements.txt
blis==0.7.4
catalogue==1.0.0
certifi==2020.12.5
chardet==4.0.0
cymem==2.0.5
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-2.3.1/en_core_web_lg-2.3.1.tar.gz
idna==2.10
murmurhash==1.0.5
numpy==1.19.5
pandas==1.2.0
plac==1.1.3
preshed==3.0.5
protobuf==3.14.0
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2020.5
requests==2.25.1
six==1.15.0
spacy==2.3.5
srsly==1.0.5
stanza==1.1.1
thinc==7.4.5
torch==1.7.1
tqdm==4.56.0
typing-extensions==3.7.4.3
urllib3==1.26.2
wasabi==0.8.0
